io: extract net_state bookkeeping from handler.c
Move three backend-agnostic tables out of lib/io/handler.c
(liburing-only) into a new lib/io/net_state.c compiled on both
backends:

  - pending_requests[] / request_mutex: in-flight outbound RPC
    tracking keyed by XID, used by the connect completion handler
    to match a CONNECT result to its originating io_send_request.

  - listener_fds[] / num_listeners: server-side listen sockets;
    io_check_for_listener_restart re-arms the accept op when a
    listener socket is closed.

  - conn_buffers[] / io_buffer_state_*: per-fd accumulating buffers
    for reassembling RPC records across partial TCP reads, plus
    io_client_fd_register / io_client_fd_unregister.

Add a new public io_net_state_fini() that both backends call from
their io_handler_fini to free the module's state after in-flight
operations have drained.  Drop the redundant memset(conn_buffers)
from io_handler_init -- static arrays are already zero-initialized.

Remove the corresponding stubs from backend_kqueue.c (nine of them).
No functional change on Linux; on FreeBSD, the kqueue backend now
has real request-table / listener / buffer-state bookkeeping instead
of silently dropping every call.

Completion handlers io_handle_read and io_handle_write are still
stubs on the kqueue backend -- porting their submission helpers
(rpc_trans_writer, io_do_tls, request_more_read_data) is PR #7.
